Route::getExactClone: read Waypoint::state low byte (uint8_t cast) to match ldrb — +1 linked_exact
ASM: original reads waypoint->state@0x130 via ldrb (byte); ours read the full uint16_t (ldrh). Casting
to uint8_t in the comparison makes the compiler emit ldrb, matching — without changing the field type
(so the strh writer stays byte-exact, no regression). parity 38/0/0, linked_exact 1903->1904.

+++ src/game/world/Route.cpp
@@ -32,7 +32,7 @@ Waypoint *Route::getWaypoint() {
 Route *Route::getExactClone() {
     Route *result = clone();
     for (uint32_t i = 0; i < result->waypoints->size(); i++) {
-        if ((*this->waypoints)[i]->state != 0)
+        if ((uint8_t) (*this->waypoints)[i]->state != 0)
             (*result->waypoints)[i]->reached();
     }
     result->currentIndex = this->currentIndex;
